optional debloat tool

+++ scripts/setup.ps1
@@ -57,8 +57,9 @@ if(!$hasArchWSL){
     Clear-Host
 }
 
-git clone "https://github.com/LeDragoX/Win-Debloat-Tools.git"
-Set-Location -Path ".\Win-Debloat-Tools"
-Set-ExecutionPolicy Unrestricted -Scope CurrentUser -Force; ls -Recurse *.ps*1 | Unblock-File; .\"Win10ScriptCLI.ps1"
-Set-Location -Path ".."
-Remove-Item -Path .\Win-Debloat-Tools -Recurse
+$runWinDebloatTools = Read-Host "Run Windows Debloat Tools script? [Y/n]"
+if($runWinDebloatTools.ToLower() -eq "y"){
+    git clone "https://github.com/LeDragoX/Win-Debloat-Tools.git"
+    Set-ExecutionPolicy Unrestricted -Scope CurrentUser -Force; ls -Recurse *.ps*1 | Unblock-File; ".\Win-Debloat-Tools\Win10ScriptCLI.ps1"
+    Remove-Item -Path .\Win-Debloat-Tools -Recurse
+}
